yarpc/pyproject.toml
[project]
name = "yarpc"
version = "0.1.0"
description = "Yet another remote procedure call generator"
authors = [
    {name = "", email = "dummy@mail.com"},
]
dependencies = [
    "jsonschema>=4.20.0",
    "Jinja2>=3.1.2",
    "pyyaml>=6.0.1",
    "setuptools>=69.0.3",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
yarpc = "yarpc.main:main"

[tool.pdm]
package-type = "application"

[tool.pdm.dev-dependencies]
dev = [
    "-e file:///${PROJECT_ROOT}/#egg=yarpc",
    "pytest>=7.4.4",
]

[tool.pdm.build.wheel-data]
data = [
    "src/yarpc/schema/*",
    "src/yarpc/templates/*",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

